Add activity_fractions() method for inventory classes, refactor test imports to fix mypy errors
